requirements.txt
aioredis==1.3.1
asgiref==3.4.1
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.12.3
Automat==20.2.0
autopep8==1.5.3
bcrypt==3.2.0
certifi==2020.6.20
cffi==1.14.5
channels==3.0.3
channels-redis==3.2.0
chardet==3.0.4
constantly==15.1.0
cryptography==3.3.1
daphne==3.0.1
Django==4.0
django-cors-headers==3.4.0
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
drf-flex-fields==0.9.0
funcy==1.16
gunicorn==20.0.4
hiredis==1.1.0
hyperlink==21.0.0
idna==2.10
incremental==17.5.0
inflection==0.3.1
msgpack==1.0.2
mysqlclient==1.4.6
paramiko==2.7.2
Pillow==8.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==2.3.0
PyNaCl==1.4.0
pyOpenSSL==20.0.1
pytz==2020.1
redis==3.5.3
requests==2.24.0
semantic-version==2.8.5
serpy==0.3.1
service-identity==18.1.0
setuptools-rust==0.11.6
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
Twisted==20.3.0
txaio==20.12.1
typing-extensions==3.10.0.0
urllib3==1.25.10
zope.interface==5.2.0
